- Bitmap ears

+++ app/app/src/main/java/net/pdev/ears/Battery.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class BatteryData : ViewModel() {
-    private val _batteryLevel: MutableLiveData<Float> by lazy {
-        MutableLiveData<Float>(0F)
+    private val _batteryLevel: MutableLiveData<UInt> by lazy {
+        MutableLiveData<UInt>(0u)
     }
 
     private val _batteryColor: MutableLiveData<UInt> by lazy {
@@ -17,11 +17,11 @@ class BatteryData : ViewModel() {
         return _batteryColor
     }
 
-    fun getBatteryLevel(): LiveData<Float> {
+    fun getBatteryLevel(): LiveData<UInt> {
         return _batteryLevel
     }
 
-    fun setBatteryLevel(level: Float) {
+    fun setBatteryLevel(level: UInt) {
         _batteryLevel.value = level
     }
 }
